+++ examples/kitchen-sink/src/pages/nav.tsx
@@ -6,7 +6,7 @@ import Link from "@twofold/framework/link";
 import { ReactNode } from "react";
 
 export default function Nav() {
-  return null;
+  // return null;
 
   let { path } = useRouter();