+++ examples/databases/CHANGELOG.md
@@ -1,5 +1,12 @@
 # databases
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @twofold/framework@0.0.20
+
 ## 0.0.12
 
 ### Patch Changes