+++ examples/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,12 @@
 # kitchen-sink
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @twofold/framework@0.0.20
+
 ## 1.0.18
 
 ### Patch Changes